Move `Toolbar` files inside `toolbar` subfolder
